mv fixtures to integration conftest
