+++ tests/integration/step-function/conftest.py
@@ -1,46 +1,9 @@
 import pytest
 import os
-import json
 import logging
 
 log = logging.getLogger(__file__)
 log.setLevel(logging.DEBUG)
-
-
-@pytest.fixture
-def mock_sf_cfg(mut_output):
-    """
-    Overwrites Step Function State Machine placeholder name with name from Terraform module.
-    See here for more info on mock config file:
-    https://docs.aws.amazon.com/step-functions/latest/dg/sfn-local-mock-cfg-file.html
-    """
-    log.info(
-        "Replacing placholder state machine name with: "
-        + mut_output["step_function_name"]
-    )
-    mock_path = os.path.join(os.path.dirname(__file__), "mock_sf_cfg.json")
-    with open(mock_path, "r") as f:
-        cfg = json.load(f)
-
-    cfg["StateMachines"][mut_output["step_function_name"]] = cfg["StateMachines"].pop(
-        "Placeholder"
-    )
-
-    with open(mock_path, "w") as f:
-        json.dump(cfg, f)
-
-    yield mock_path
-
-    log.info("Replacing state machine name back with placholder")
-    with open(mock_path, "r") as f:
-        cfg = json.load(f)
-
-    cfg["StateMachines"]["Placeholder"] = cfg["StateMachines"].pop(
-        mut_output["step_function_name"]
-    )
-
-    with open(mock_path, "w") as f:
-        json.dump(cfg, f, indent=4, sort_keys=True)
 
 
 @pytest.fixture
